1.0.4 update

+++ src/CodeGenerator.ts
@@ -1,11 +1,10 @@
 import * as vscode from 'vscode';
-import * as mysql from 'mysql';
 import * as nunjucks from 'nunjucks';
 import * as path from 'path';
 import * as fs from 'fs';
 import { CodeEntity } from './CodeEntity';
 import { GycTools } from './GycTools';
-import { columnQuerierFactory, ColumnQuerier } from './DataBase';
+import { columnQuerierFactory } from './DataBase';
 
 export class CodeGenerator {
 
@@ -18,27 +17,27 @@ export class CodeGenerator {
     sourceDatabaseConfig: GycTools.DatabaseConfig;
 
 
-    readonly columnQuerier:ColumnQuerier;
+    readonly columnQuerier: GycTools.ColumnQuerier;
 
     constructor(context: vscode.ExtensionContext, selectedItem: any, targetProject: GycTools.ProjectConfig) {
         this.context = context;
         this.selectedTableName = selectedItem.value;
         this.targetProject = targetProject;
-        this.columnQuerier=columnQuerierFactory(selectedItem);
+        this.columnQuerier = columnQuerierFactory(selectedItem);
         this.sourceDatabaseConfig = this.getSourceDatabase();
     }
 
-    generateNew(){
-        this.columnQuerier.getTableColumnInfo().then((columnArray)=>{
+    generateNew() {
+        this.columnQuerier.getTableColumnInfo().then((columnArray) => {
             let codeEntity = new CodeEntity(this.columnQuerier, columnArray, this.targetProject.baseModelProperties, this.sourceDatabaseConfig);
 
-            const templatePath = GycTools.Utils.getTemplatePath(this.context,this.targetProject);
+            const templatePath = GycTools.Utils.getTemplatePath(this.context, this.targetProject);
             if (!fs.existsSync(templatePath)) {
                 vscode.window.showErrorMessage('template path error');
                 return;
             }
 
-            nunjucks.configure(templatePath, { autoescape: false,trimBlocks:true,lstripBlocks:true }); //config template path
+            nunjucks.configure(templatePath, { autoescape: false, trimBlocks: true, lstripBlocks: true }); //config template path
             let temlpateList: any[] = this.sourceDatabaseConfig.templateList; //get templates list config
             temlpateList = temlpateList.filter((e) => { return e.enabled === undefined || e.enabled === true; });
             temlpateList.forEach((t) => {
@@ -50,17 +49,21 @@ export class CodeGenerator {
                 }
                 const outFliePath = path.join(outDirectory, `${codeEntity.className}` + t.outFileType);//generate code file
                 if (renderString.length > 0) {
-                    fs.writeFileSync(outFliePath, renderString);
-                    if (this.targetProject.openFileWhenComplete) {
-                        const doc = vscode.workspace.openTextDocument(outFliePath).then((d) => {
-                            vscode.window.showTextDocument(d, { preserveFocus: false, viewColumn: vscode.ViewColumn.Beside });
-                        });
+                    if (fs.existsSync(outFliePath) && this.targetProject.overrideExists !== undefined && this.targetProject.overrideExists === false) {
+                        vscode.window.showInformationMessage(`${codeEntity.className}` + t.outFileType + ' file exists');
+                    } else {
+                        fs.writeFileSync(outFliePath, renderString);
+                        if (this.targetProject.openFileWhenComplete) {
+                            const doc = vscode.workspace.openTextDocument(outFliePath).then((d) => {
+                                vscode.window.showTextDocument(d, { preserveFocus: false, viewColumn: vscode.ViewColumn.Beside });
+                            });
+                        }
                     }
                 } else {
                     vscode.window.showErrorMessage('template error:', t.templateName);
                 }
             });
-            vscode.window.showInformationMessage('gyctools:generate successfully','OK');
+            vscode.window.showInformationMessage('gyctools:generate successfully', 'OK');
 
         }).catch((error) => {
             console.error(error);
